Adding shots and stations as I'd like them to be
- This is a big ugly commit but represents a 
spike on the domain.

+++ lib/compassIO/DatFile/shot.ex
@@ -31,7 +31,7 @@ defmodule CompassIO.DatFile.Shot do
   end
 
   def azimuth(shot) do
-    {val, _} = Integer.parse(shot.bearing)
+    {val, _} = Float.parse(shot.bearing)
     val
   end
